Add checksum validation and split into several files
